Fulfilled "Adding Debits" user story
Used a form that keeps track of the user input as it comes in. On
submit, a handler is called that finishes creating the object for the
debit. The object is designed to be the same as the objects from the API
call so that the mapping function runs smoothly.