Fixed persistence of debit and credit data
There was a problem having the use effect for the fetch call in the
credit and debit components. Every time the component loaded, it would
reset the state to have the values from the API. A simple fix to this
was moving the debit and credit states into the App component along with
the fetch calls. Both fetch calls are now in the App component under a
use effect hook with an empty dependency array--so it only runs once
when the user loads on to the website.