Fixed duplicaty key issue when adding entries
There was a problem where if a user pressed the add credit/debit option
multiple times, the same key id would be used for those entries. This
was because unique IDs were only generated when the input boxes were
changed. I did consider resetting the text boxes, but that wouldn't
solve the root problem since the user can currently add empty values.
More importantly, the user may legitamtely want to add multiple debit
and credit values if they want to add more than one of the same item.

Thankfully, this was simply fixed by only generating the id when the
user clicks the add button. The generation of the date has also been
moved to the add button handler as well. This saves on code executions
and more importantly, ensures that each item has a unique key value for
React to use.